Move color trait type to the widget package
